test: move out spec tests
